Added score counter and return to the menu button on the game. Removed the settings button from the main menu and also created the Coroutine script to get the set sail sound play when pressing set sail button (I was not successful)
Added score counter and return to the menu button on the game. Removed the settings button from the main menu and also created the Coroutine script to get the set sail sound play when pressing set sail button (I was not successful). I already added two game objects to the main menu scene: Audio which has the sound file (It is already connected to the set sail button on click component.)
Timer has Coroutine script connected to it, but it is not yet connected to any buttons